NXDRIVE-2112: Always start a fresh download on Direct Edit
It seems that there are more and more issues because of a previous
Direct Edit download failure. Doing another Direct Edit on the same
document will resume the download. Sometimes it will work, and
most of the time not.

A fresh download is now started everytime to prevent this issues.
The test test_resumed_download has benn added.
Also changelog has been updated.